% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{algarawi}
\alias{algarawi}
\title{Top table of changes with gestation in maternal blood by Al-Grawi et al.
A dataset containing the genes specifically expressed in different type of cells from the 
paper PMID: 27711190}
\format{A data frame with 32 rows and 8 variables:
\describe{
  \item{ID}{Illumina probe ID}
  \item{adj.P.Val}{Adjusted p-value}
 \item{logFC}{log2 fold change betwene late and early gestation}
  \item{P.Value}{Nominal p-value}
  \item{Gene.symbol}{Gene symbol}
}}
\source{
\url{}
}
\usage{
algarawi
}
\description{
Top table of changes with gestation in maternal blood by Al-Grawi et al.
A dataset containing the genes specifically expressed in different type of cells from the 
paper PMID: 27711190
}
\keyword{datasets}
