% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ano32}
\alias{ano32}
\title{Annotation data for samples
A dataset containing the annotation of 32 maternal blood samples}
\format{A data frame with 32 rows and 8 variables:
\describe{
  \item{SampleID}{Unique sample identifier}
  \item{IndividualID}{Individual ID}
  \item{GA}{Gestational Age, in weeks}
  \item{Group}{Labor Group: TIL term in labor; TNL term not in labor}
  \item{RIN}{RNA Integrity Number }
  \item{StorageMonths}{Sample storage duration from collection to analysis in months}
  \item{GAAnalysis}{Sample was used in the GA effect analysis}
  \item{LaborAnalysis}{Sample was used in the labor effect analysis}
}}
\source{
\url{}
}
\usage{
ano32
}
\description{
Annotation data for samples
A dataset containing the annotation of 32 maternal blood samples
}
\keyword{datasets}
