% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ccount}
\alias{Ccount}
\title{Gene expression count matrix obtained with DriverMap (Cellecta Inc)
Salmon quantified gene expression for 32 maternal blood samples}
\format{A matrix of counts of sequencing fragments
\describe{
rows are ENSEMBLE gene identifiers and columns are samples matching rows in ano32
}}
\source{
\url{}
}
\usage{
Ccount
}
\description{
Gene expression count matrix obtained with DriverMap (Cellecta Inc)
Salmon quantified gene expression for 32 maternal blood samples
}
\keyword{datasets}
