% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esetHTA}
\alias{esetHTA}
\title{Gene expression matrix obtained with Affymetrix HTA 2.0 arrays
HTA gene expression for 32 maternal blood samples}
\format{A matrix with RMA log2 expression summaries
\describe{
rows are transcript cluster IDs and columns are samples matching rows in ano32
}}
\source{
\url{}
}
\usage{
esetHTA
}
\description{
Gene expression matrix obtained with Affymetrix HTA 2.0 arrays
HTA gene expression for 32 maternal blood samples
}
\keyword{datasets}
