% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esetPCR}
\alias{esetPCR}
\title{Gene expression matrix obtained with qRT-PCR
qRT-PCR gene expression for 32 maternal blood samples}
\format{A matrix with -Delta CT values
\describe{
rows are gene symbols and columns are samples matching rows in ano32
}}
\source{
\url{}
}
\usage{
esetPCR
}
\description{
Gene expression matrix obtained with qRT-PCR
qRT-PCR gene expression for 32 maternal blood samples
}
\keyword{datasets}
