% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heng}
\alias{heng}
\title{Top table of changes with gestation in maternal blood by Heng et al.
A dataset containing the genes specifically expressed in different type of cells from the 
paper PMID: 27333071}
\format{A data frame with 32 rows and 8 variables:
\describe{
  \item{SYMBOL}{Gene symbol}
  \item{logFC}{log2 fold change betwene late and early gestation}
  \item{P.Value}{Nominal p-value}
  \item{adj.P.Val}{Adjusted p-value}
  \item{Gene.symbol}{Gene symbol}
}}
\source{
\url{}
}
\usage{
heng
}
\description{
Top table of changes with gestation in maternal blood by Heng et al.
A dataset containing the genes specifically expressed in different type of cells from the 
paper PMID: 27333071
}
\keyword{datasets}
