% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npspge}
\alias{npspge}
\title{Number of expressed exon level probesets per transcript cluster
Obtained based on detection above background p-values from TAC from cell files corresponding to esetHTA}
\format{A data frame with 32 rows and 8 variables:
\describe{
  \item{npspge}{number of probesets expressed above background in 5 or more of the 32 samples. Names are transcript cluster IDS}
}}
\source{
\url{}
}
\usage{
npspge
}
\description{
Number of expressed exon level probesets per transcript cluster
Obtained based on detection above background p-values from TAC from cell files corresponding to esetHTA
}
\keyword{datasets}
