% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCGeneSets}
\alias{SCGeneSets}
\title{Single cell transcripomics gene signatures
A dataset containing the genes specifically expressed in different type of cells from the 
paper PMID: 28830992}
\format{A data frame with 32 rows and 8 variables:
\describe{
  \item{Symbol}{Gene Symbol}
  \item{Type}{Cell type}
}}
\source{
\url{}
}
\usage{
SCGeneSets
}
\description{
Single cell transcripomics gene signatures
A dataset containing the genes specifically expressed in different type of cells from the 
paper PMID: 28830992
}
\keyword{datasets}
